Update main.py
decided py-script wouldn't be a good fit, so decided to try making a model in flask. Today I stopped experimenting with py-script and decided to make a sample model that I can use with flask. This is a model that I created that predicts if a person is likely to have heart disease